Created new class files and updated MainActivity
The user can now register their information using the createProfile button on MainActivity.
User information is saved and is stored into a newly created database.
Upon startup of app, the user will always be prompted to enter their credentials, even if they have already registered their information. This will be worked on in future updates.
Would like to implement a database checker to check if database and user already exist. If user exists, the app will start on their overview page and not on the registration page. This will be worked on in future updates.